python/requirements.txt
pytz==2013d
jenkinsapi==0.2.13
pyserial==2.6
